+++ classes/Controllers/Compatibility.php
@@ -28,6 +28,7 @@ class Compatibility extends Controller {
 	public function init() {
 		$this->container->register_controllers( [
 			'Compatibility\Backcompat\Filters'     => new \PopupMaker\Controllers\Compatibility\Backcompat\Filters( $this->container ),
+			'Compatibility\Builder\Astra'          => new \PopupMaker\Controllers\Compatibility\Builder\Astra( $this->container ),
 			'Compatibility\SEO\Yoast'              => new \PopupMaker\Controllers\Compatibility\SEO\Yoast( $this->container ),
 			'Compatibility\Builder\Divi'           => new \PopupMaker\Controllers\Compatibility\Builder\Divi( $this->container ),
 			'Compatibility\Plugin\ACF'             => new \PopupMaker\Controllers\Compatibility\Plugin\ACF( $this->container ),